a serious bug for composite shell quartet code
we found a serious bug for composite shell quartet, for example; the
SP shell quartet. The result abcd[index] the index calculating from
previous way is not correct.

Now we correct it, and revise the testing code. The previous testing
code is also not correct. The testing code is also revised.

The result is not tested yet